requirements.txt
discord.py>=2.3.0
groq>=0.9.0
python-dotenv>=1.0.0
aiosqlite>=0.19.0